refactor: reorganize project structure to standard src layout
