+++ src/hourei_api/__init__.py
@@ -1,0 +1,19 @@
+from .hourei_apiv2 import (
+    get_lawid_from_lawtitle,
+    get_lawdata_from_law_id,
+    get_lawdata_from_lawname,
+    save_xml_string_to_file,
+    extract_sections_from_xml,
+)
+from .text_converter import convert_xml_to_text, LawXmlParser
+from .yaml_converter import convert_xml_to_yaml, LawToYamlConverter
+
+# LLM/LangGraph機能はオプショナル依存関係のため、インストールされていない場合は無視する
+try:
+    from .law_extraction_v2 import (
+        LegalExtractionConfig,
+        create_legal_extraction_system,
+        YamlArticleExtractor,
+    )
+except ImportError:
+    pass
